requirements.txt
altgraph==0.17.5
appdirs==1.4.4
argcomplete==3.6.3
asn1crypto==1.5.1
attrs==24.2.0
Babel==2.10.3
bcrypt==4.1.2
beautifulsoup4==4.12.3
cbor2==5.6.2
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==42.0.8
decorator==5.1.1
docopt==0.6.2
docutils==0.20.1
et-xmlfile==1.1.0
freezegun==1.2.1
geoip2==2.9.0
idna==3.11
importlib_resources==6.5.2
isodate==0.6.1
Jinja2==3.1.2
libsass==0.22.0
lxml==5.2.1
lxml_html_clean==0.2.2
MarkupSafe==2.1.5
maxminddb==2.6.2
num2words==0.5.13
ofxparse==0.21
openpyxl==3.1.2
packaging==25.0
passlib==1.7.4
pefile==2024.8.26
pillow==12.0.0
platformdirs==4.3.6
polib==1.1.1
psutil==5.9.8
psycopg2==2.9.9
pycparser==2.23
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
pyOpenSSL==24.1.0
PyPDF2==2.12.1
pypiwin32==223
pypng==0.20220715.0
pyserial==3.5
python-barcode==0.16.1
python-dateutil==2.8.2
python-escpos==3.1
python-stdnum==1.19
pytz==2024.2
pyusb==1.2.1
pywin32==307
pywin32-ctypes==0.2.3
PyYAML==6.0.3
qrcode==8.2
reportlab==4.1.0
requests==2.32.5
requests-file==2.1.0
requests-toolbelt==1.0.0
rjsmin==1.2.0
rl-renderPM==4.0.3
setuptools==80.9.0
six==1.17.0
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.5.0
vobject==0.9.6.1
Werkzeug==3.0.1
wheel==0.44.0
xlrd==2.0.1
XlsxWriter==3.1.9
xlwt==1.3.0
zeep==4.2.1
